+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# HEAD
+
+## Bug Fixes
+
+-   **js:** Enable passing `samesite` when calling removeItem() (#2)
+
 # 1.0.1
 
 ## Bug Fixes
@@ -5,6 +11,7 @@
 -   **node:** Update version number after accidental publish before a pull.
 
 # 1.0.0
+
 ## Bug Fixes
 
 -   **js:** Remove `Mozilla` JS namespace from library.